Make all the C files in the libraries compile under MSVC 2005 Express.  There are still a few warnings, and probably some subtle issues, but it's better than nothing.
svn:r499

+++ WIN32-Code/misc.c
@@ -25,13 +25,13 @@
 
 #ifndef HAVE_GETTIMEOFDAY
 int gettimeofday(struct timeval *tv, struct timezone *tz) {
-  struct _timeb tb;
+	struct _timeb tb;
 
 	if(tv == NULL)
 		return -1;
 
 	_ftime(&tb);
-	tv->tv_sec = tb.time;
+	tv->tv_sec = (long) tb.time;
 	tv->tv_usec = ((int) tb.millitm) * 1000;
 	return 0;
 }
